requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
colorama==0.4.3
distlib==0.3.1
dj-database-url==0.5.0
Django==3.0.8
django-crum==0.7.6
django-heroku==0.3.1
django-widget-tweaks==1.4.8
filelock==3.0.12
gunicorn==20.0.4
importlib-metadata==1.7.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.2.0
psycopg2==2.8.5
pylint==2.5.3
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
virtualenv==20.0.25
whitenoise==5.1.0
wrapt==1.12.1
zipp==3.1.0
